sdasd
Hi, this is a simple hand written digit recognition program. I do not know why I only upload the dumbest shit to my GitHub but this is genuinely useful to no one. This was commented out by ChatGPT btw because I hate commenting(bad habit do not follow suite).

Be sure to pip install everything
I am not giving a requirements.txt because lol if you're reading this you're at the lowest point of your career

make sure you have the .png images saved in a folder named 'digits' and that all the images are named as digit1, digit2, digit3... etc.

simple rerun the program to go to the next image

There is a portion that you must uncomment if you want to only run the saved model

so be sure to check that out lol

increase the number of epochs if it isnt that accurate

but anything 99 accuracy above is enough you greedy person

if you've gotten to the end of this

rethink your life and give me a follow
maybe even message me